+++ features/step-definitions/player-movement.steps.js
@@ -31,6 +31,10 @@ Given('stairs at position {int}, {int}', function (x, y) {
   this.map.setTile(x, y, TILE.STAIRS_DOWN);
 });
 
+Given('a door at position {int}, {int}', function (x, y) {
+  this.map.setTile(x, y, TILE.DOOR);
+});
+
 // --- When ---
 
 When('the player moves right', function () {